Run charmhub-lp-tool when merging pull requests.
This change introduces a new job called "dry-run" which runs
charmhub-lp-tool sync and ensure-series in anonymous mode on pull
requests to capture any possible issues.

To apply the changes merged the github workflow deploy.yaml run
charmhub-lp-tool sync and ensure-series with the --i-really-mean-it
flag on the 'push' event, this workflow only runs when the github repo
is openstack-charmers/charmed-openstack-info and the branch is 'main'.

The logs are uploaded as artifact for inspection.

The token used belongs to the uosci-testing-bot user[0]

[0] https://launchpad.net/~uosci-testing-bot