Add infinidat charms
This patch adds the lp-builder-config metadata for the 3 infinidat
charms (cinder-infinidat, manila-infinidat and infinidat-tools) that
support the Infinidat storage backend on OpenStack.  They've been
put in a separate infinidat.yaml file as they only start from ussuri and
thus the entire branches config would need to be replicated 3 times if
it was included in the openstack.yaml.  This way, it's a bit cleaner
until there is a better inheritance system in charmed-openstack-info
lp-builder-config.