Modify the infinidat charms
This changes the infinidat charms CI/CD metadata so that:

 - stable/ussuri channel pushes to ussuri, victoria, wallaby and
   xena
 - stable/yoga channel pushes to yoga for focal and jammy
 - stable/zed pushs to zed, 2023.1 for jammy

This is because it is essentially the same charm, and they have just
been split by LTS, but the yoga charm still needs to support a series
upgrade from focal -> jammy.  The zed+ charm then supports jammy only.
This allows additional branches to be inserted as needed, but it is felt
unlikely that this will be needed.

The latest channel is for development, and supports jammy only.