Add bobcat 2023.2/candidate channels
Add the bobcat 2023.2 channes for the OpenStack charms. The recipe is to
push to /candidate for SQA to consume prior to releasing to stable. Bug
fixes during the testing period will therefore be pushed to /candidate
to assist with simplified testing of any fixes.

The charmcraft 2.2/stable channel has been selected for the recipes as
that is the latest stable channel availble for long term building of the
bobcat charms.

Bases 22.04 and 23.10 have been selected to support jammy and mantic.