Remove unknown recipes when syncing
The lp-build-config has to be the last word in recipes for the repos/LP
projects that are managed as if a channel is moved from one branch to
another the *old* recipe will still be maintained resulting in two
recipes pushing to the same branch. This could lead to really weird
conflicts.

Thus, either charmed-openstack-info *is* the only metadata and controls
the repos automatically, or we need to remove the auto-syncing.  I'm in
favour of the former, and so this change removes unknown (to
charmed-openstack-info) recipes when merging a change to the repos.